requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
pyTelegramBotAPI==3.6.6
requests==2.21.0
six==1.12.0
urllib3==1.24.2
Flask==1.1.1
python-crontab==2.3.8

